Updated info under experience
var descriptions: Performing case management work and create account for new employee. The NHS login project, started in August 2021 – April 2022 where I was contracted as a support annalist.  In this role I used a number of applications on a daily basis such as: Splunk, Cherwell, Jira, confluence, all used within a VDI. I had to investigate possible errors in logs and based on the data collected either send the user a resolution as to how to resolve the issue or escalate it to the right team. I also had to manage different channel weekly to look at particular cases. I was also responsible for opening and closing the store and reported directly to the store owner. Personal responsibility was a key skill that I learned from this role.
var internType: "Junior DevSecOps Engineer"